// types/graphql-types.ts
export type CreatePagesFromMicroCmsQuery = { allMicrocmsPosts: { edges: Array<{ node: { id: string, description?: string | null, postsId?: string | null, publishedAt?: any | null, publishedAtOnHatena?: any | null, title?: string | null, internal: { type: string }, tags?: Array<{ color?: string | null, id?: string | null, name?: string | null } | null> | null, thumbnail?: { url?: string | null } | null } }> }, allFeedQiitaPost: { edges: Array<{ node: { contentSnippet?: string | null, id: string, isoDate?: any | null, link?: string | null, title?: string | null, internal: { type: string } } }> }, allFeedZennPost: { edges: Array<{ node: { contentSnippet?: string | null, id: string, isoDate?: any | null, link?: string | null, title?: string | null, internal: { type: string } } }> } };

export type CreateTagsPagesFromMicroCmsQueryVariables = Exact<{ [key: string]: never; }>;


export type CreateTagsPagesFromMicroCmsQuery = { allMicrocmsTags: { edges: Array<{ node: { name?: string | null, tagsId?: string | null } }> } };

export type CreateTagPostsPagesFromMicroCmsQueryVariables = Exact<{
  tagsId?: InputMaybe<Scalars['String']>;
}>;


export type CreateTagPostsPagesFromMicroCmsQuery = { allMicrocmsPosts: { totalCount: number } };